Create "Value" Component and Inject HTTP Client
Created Value component and used dependency
injection to inject the HttpClient in the constructor.
Wrote getValues method to make a GET request
to our Web API and store our values in a property
called values. Displayed these values in the
HTML file for the component. Used the ngOnInit
lifecycle method to call the getValues method.