Create Home and Register Components
Created HomeComponent for the landing
page and created the RegisterComponent
to be a child component of it. Clicking the
Register button brings up a form for new
user registration. Injected AuthService into
the constructor of the RegisterComponent
and wrote a register method to call the
HTTP POST method of register in AuthService.